refactoring part of staff tests
