+++ .travis.yml
@@ -6,7 +6,7 @@ install:
   - pip install -r requirements.txt
   - pip install coveralls
 script:
-  - coverage run manage.py test evap.evaluation evap.staff evap.contributor evap.results evap.student evap.grades evap.rewards evap.staff.tests.TestDataTest.load_test_data
+  - coverage run manage.py test evap.evaluation evap.staff evap.contributor evap.results evap.student evap.grades evap.rewards evap.staff.tests_general.TestDataTest.load_test_data
 after_success:
   - coveralls
 sudo: false
