+++ evap/contributor/tests/test_forms.py
@@ -5,8 +5,8 @@ from model_mommy import mommy
 
 from evap.evaluation.models import UserProfile, Course, Questionnaire, Contribution
 from evap.contributor.forms import DelegatesForm, EditorContributionForm
+from evap.evaluation.tests.test_utils import get_form_data_from_instance
 from evap.staff.forms import ContributionFormSet
-from evap.staff.tests import get_form_data_from_instance
 
 
 class UserFormTests(TestCase):
